update: 更新文件 avrouter_impl.cpp

+++ services/session/server/avrouter_impl.cpp
@@ -816,18 +816,18 @@ void AVRouterImpl::SendCommandArgsToCast(const int64_t castHandle, const int32_t
     const std::string& params)
 {
     SLOGI("AVRouterImpl start send commandArgs to cast");
-
-    // The first 32 bits are providerId, the last 32 bits are castId
-    int32_t providerNumber = static_cast<uint64_t>(static_cast<uint64_t>(castHandle)) >> 32;
+ 
+    // the first 32 bits are providerId, the last 32 bits are castId
+    int32_t providerNumber = static_cast<int32_t>(static_cast<const uint64_t>(castHandle) >> 32);
     SLOGI("Get hwcastprovider of provider %{public}d", providerNumber);
-
-    int32_t castId = static_cast<int32_t>((static_cast<uint64_t>(castHandle) << 32) >> 32);
-
+ 
+    int32_t castId = static_cast<int32_t>((static_cast<const uint64_t>(castHandle) << 32) >> 32);
+ 
     CHECK_AND_RETURN_LOG(providerManagerMap_.find(providerNumber) != providerManagerMap_.end(),
         "Can not find corresponding provider");
     CHECK_AND_RETURN_LOG(providerManagerMap_[providerNumber] != nullptr && providerManagerMap_[providerNumber]->
         provider_ != nullptr, "provider is nullptr");
-
+ 
     providerManagerMap_[providerNumber]->provider_->SendCommandArgsToCast(castId, commandType, params);
 }
